See merge request trixnity/trixnity!369

Use long for powerlevels

the spec allows more than int

+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/user/UserService.kt
@@ -48,12 +48,12 @@ interface UserService {
 
     fun getReceiptsById(roomId: RoomId, userId: UserId): Flow<RoomUserReceipts?>
 
-    fun getPowerLevel(roomId: RoomId, userId: UserId): Flow<Int>
+    fun getPowerLevel(roomId: RoomId, userId: UserId): Flow<Long>
     fun getPowerLevel(
         userId: UserId,
         powerLevelsEventContent: PowerLevelsEventContent?,
         createEventContent: CreateEventContent
-    ): Int
+    ): Long
 
     fun canKickUser(roomId: RoomId, userId: UserId): Flow<Boolean>
     fun canBanUser(roomId: RoomId, userId: UserId): Flow<Boolean>
@@ -62,7 +62,7 @@ interface UserService {
     fun canInvite(roomId: RoomId): Flow<Boolean>
     fun canRedactEvent(roomId: RoomId, eventId: EventId): Flow<Boolean>
 
-    fun canSetPowerLevelToMax(roomId: RoomId, userId: UserId): Flow<Int?>
+    fun canSetPowerLevelToMax(roomId: RoomId, userId: UserId): Flow<Long?>
 
     fun canSendEvent(roomId: RoomId, eventClass: KClass<out EventContent>): Flow<Boolean>
 
@@ -143,7 +143,7 @@ class UserServiceImpl(
     override fun getPowerLevel(
         roomId: RoomId,
         userId: UserId
-    ): Flow<Int> =
+    ): Flow<Long> =
         combine(
             roomStateStore.getContentByStateKey<PowerLevelsEventContent>(roomId),
             roomStateStore.getContentByStateKey<CreateEventContent>(roomId).filterNotNull()
@@ -155,7 +155,7 @@ class UserServiceImpl(
         userId: UserId,
         powerLevelsEventContent: PowerLevelsEventContent?,
         createEventContent: CreateEventContent
-    ): Int {
+    ): Long {
         return when (powerLevelsEventContent) {
             null -> if (createEventContent.creator == userId) 100 else 0
             else -> powerLevelsEventContent.users[userId] ?: powerLevelsEventContent.usersDefault
@@ -263,7 +263,7 @@ class UserServiceImpl(
     override fun canSetPowerLevelToMax(
         roomId: RoomId,
         userId: UserId
-    ): Flow<Int?> {
+    ): Flow<Long?> {
         return combine(
             getPowerLevel(roomId, ownUserId),
             getPowerLevel(roomId, userId),
