+++ trixnity-client/src/commonTest/kotlin/net/folivo/trixnity/client/mocks/UserServiceMock.kt
@@ -82,7 +82,7 @@ class UserServiceMock : UserService {
         TODO("Not yet implemented")
     }
 
-    override fun getPowerLevel(roomId: RoomId, userId: UserId): Flow<Int> {
+    override fun getPowerLevel(roomId: RoomId, userId: UserId): Flow<Long> {
         throw NotImplementedError()
     }
 
@@ -90,14 +90,14 @@ class UserServiceMock : UserService {
         userId: UserId,
         powerLevelsEventContent: PowerLevelsEventContent?,
         createEventContent: CreateEventContent
-    ): Int {
+    ): Long {
         throw NotImplementedError()
     }
 
     override fun canSetPowerLevelToMax(
         roomId: RoomId,
         userId: UserId
-    ): Flow<Int?> {
+    ): Flow<Long?> {
         throw NotImplementedError()
     }
